Moved LauncherActivity.kt and MainActivity.kt in the package 'ui/activity'. Also affected activity_launcher.xml and activity_main.xml as well as AndroidManifest.xml.
