+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
 		<!--~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Activities ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~-->
 
 		<activity
-			android:name = ".ui.launcher.LauncherActivity">
+			android:name = ".ui.activity.LauncherActivity">
 
 			<intent-filter>
 
@@ -34,7 +34,7 @@
 		</activity>
 
 		<activity
-			android:name = ".ui.main.MainActivity" />
+			android:name = ".ui.activity.MainActivity" />
 
 	</application>
