+++ app/src/main/res/layout/activity_launcher.xml
@@ -5,7 +5,7 @@
 	xmlns:tools = "http://schemas.android.com/tools"
 	android:layout_width = "match_parent"
 	android:layout_height = "match_parent"
-	tools:context = ".ui.launcher.LauncherActivity">
+	tools:context = ".ui.activity.LauncherActivity">
 
 	<ImageView
 		android:layout_width = "wrap_content"
